Moving documents to a separate directory (#4153)
Moving documents and it's actions, reducers to a separate directory and fixing tests
