+++ client/app/reader/DocTypeColumn.jsx
@@ -1,10 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-
-import { selectCurrentPdfLocally } from '../reader/PdfViewer/PdfViewerActions';
-
 import { bindActionCreators } from 'redux';
+import { selectCurrentPdfLocally } from '../reader/Documents/DocumentsActions';
+
 import Link from '../components/Link';
 import Highlight from '../components/Highlight';
