+++ client/app/reader/Documents/DocumentsReducer.js
@@ -1,7 +1,8 @@
 import _ from 'lodash';
 
-import * as Constants from '../constants';
+import * as Constants from '../Documents/actionTypes';
 import { update } from '../../util/ReducerUtil';
+import { DOCUMENTS_OR_COMMENTS_ENUM } from '../constants';
 
 export const initialState = {};
 
@@ -38,7 +39,7 @@ const documentsReducer = (state = initialState, action = {}) => {
   case Constants.SET_VIEWING_DOCUMENTS_OR_COMMENTS:
     return _.mapValues(state, (doc) => ({
       ...doc,
-      listComments: action.payload.documentsOrComments === Constants.DOCUMENTS_OR_COMMENTS_ENUM.COMMENTS
+      listComments: action.payload.documentsOrComments === DOCUMENTS_OR_COMMENTS_ENUM.COMMENTS
     }));
   case Constants.TOGGLE_COMMENT_LIST:
     return update(state, {
