+++ client/app/reader/PdfViewer/PdfViewerActions.js
@@ -4,7 +4,6 @@ import * as Constants from '../constants';
 import ApiUtil from '../../util/ApiUtil';
 import { CATEGORIES, ENDPOINT_NAMES } from '../analytics';
 import { selectAnnotation } from '../../reader/AnnotationLayer/AnnotationActions';
-import { hideErrorMessage, showErrorMessage } from '../commonActions';
 
 export const collectAllTags = (documents) => ({
   type: Constants.COLLECT_ALL_TAGS_FOR_OPTIONS,
@@ -60,104 +59,6 @@ export const setDocScrollPosition = (scrollTop) => ({
     scrollTop
   }
 });
-
-/** Tags **/
-
-export const newTagRequestSuccess = (docId, createdTags) =>
-  (dispatch, getState) => {
-    dispatch({
-      type: Constants.REQUEST_NEW_TAG_CREATION_SUCCESS,
-      payload: {
-        docId,
-        createdTags
-      }
-    });
-    const { documents } = getState().readerReducer;
-
-    dispatch(collectAllTags(documents));
-  }
-;
-
-export const newTagRequestFailed = (docId, tagsThatWereAttemptedToBeCreated) => (dispatch) => {
-  dispatch(showErrorMessage('tag'));
-  dispatch({
-    type: Constants.REQUEST_NEW_TAG_CREATION_FAILURE,
-    payload: {
-      docId,
-      tagsThatWereAttemptedToBeCreated
-    }
-  });
-};
-
-export const removeTagRequestFailure = (docId, tagId) => (dispatch) => {
-  dispatch(showErrorMessage('tag'));
-  dispatch({
-    type: Constants.REQUEST_REMOVE_TAG_FAILURE,
-    payload: {
-      docId,
-      tagId
-    }
-  });
-};
-
-export const removeTagRequestSuccess = (docId, tagId) =>
-  (dispatch, getState) => {
-    dispatch(hideErrorMessage('tag'));
-    dispatch({
-      type: Constants.REQUEST_REMOVE_TAG_SUCCESS,
-      payload: {
-        docId,
-        tagId
-      }
-    });
-    const { documents } = getState().readerReducer;
-
-    dispatch(collectAllTags(documents));
-  };
-
-export const removeTag = (doc, tagId) =>
-  (dispatch) => {
-    dispatch({
-      type: Constants.REQUEST_REMOVE_TAG,
-      payload: {
-        docId: doc.id,
-        tagId
-      }
-    });
-    ApiUtil.delete(`/document/${doc.id}/tag/${tagId}`, {}, ENDPOINT_NAMES.TAG).
-      then(() => {
-        dispatch(removeTagRequestSuccess(doc.id, tagId));
-      }, () => {
-        dispatch(removeTagRequestFailure(doc.id, tagId));
-      });
-  };
-
-export const addNewTag = (doc, tags) =>
-  (dispatch) => {
-    const currentTags = doc.tags;
-
-    const newTags = _(tags).
-      differenceWith(currentTags, (tag, currentTag) => tag.value === currentTag.text).
-      map((tag) => ({ text: tag.label })).
-      value();
-
-    if (_.size(newTags)) {
-      dispatch(hideErrorMessage('tag'));
-      dispatch({
-        type: Constants.REQUEST_NEW_TAG_CREATION,
-        payload: {
-          newTags,
-          docId: doc.id
-        }
-      });
-      ApiUtil.post(`/document/${doc.id}/tag`, { data: { tags: newTags } }, ENDPOINT_NAMES.TAG).
-        then((data) => {
-          dispatch(newTagRequestSuccess(doc.id, data.body.tags));
-        }, () => {
-          dispatch(newTagRequestFailed(doc.id, newTags));
-        });
-    }
-  };
 
 /** Getting Appeal Details **/
 
@@ -228,26 +129,3 @@ export const hideSearchBar = () => ({
 export const resetSidebarErrors = () => ({
   type: Constants.RESET_PDF_SIDEBAR_ERRORS
 });
-
-/** Set current PDF **/
-export const selectCurrentPdfLocally = (docId) => (dispatch) => {
-  dispatch(handleSetLastRead(docId));
-  dispatch({
-    type: Constants.SELECT_CURRENT_VIEWER_PDF,
-    payload: {
-      docId
-    }
-  });
-};
-
-export const selectCurrentPdf = (docId) => (dispatch) => {
-  ApiUtil.patch(`/document/${docId}/mark-as-read`, {}, ENDPOINT_NAMES.MARK_DOC_AS_READ).
-    catch((err) => {
-      // eslint-disable-next-line no-console
-      console.log('Error marking as read', docId, err);
-    });
-
-  dispatch(
-    selectCurrentPdfLocally(docId)
-  );
-};
