+++ client/app/reader/SideBarIssueTags.jsx
@@ -5,7 +5,7 @@ import _ from 'lodash';
 
 import CannotSaveAlert from '../reader/CannotSaveAlert';
 import SearchableDropdown from '../components/SearchableDropdown';
-import { addNewTag, removeTag } from '../reader/PdfViewer/PdfViewerActions';
+import { addNewTag, removeTag } from '../reader/Documents/DocumentsActions';
 
 class SideBarIssueTags extends PureComponent {
   render() {
