+++ client/app/reader/commonActions.js
@@ -3,9 +3,12 @@ import { getUpdatedFilteredResults } from './searchFilters';
 import {
   HIDE_ERROR_MESSAGE,
   SHOW_ERROR_MESSAGE,
-  UPDATE_FILTERED_RESULTS,
-  ASSIGN_DOCUMENTS
+  UPDATE_FILTERED_RESULTS
 } from './constants';
+
+import {
+  ASSIGN_DOCUMENTS
+} from './Documents/actionTypes';
 
 // errors
