+++ client/test/node/reader/DecisionReviewer-test.js
@@ -20,9 +20,9 @@ import { formatDateStr } from '../../../app/util/DateUtil';
 import readerReducer from '../../../app/reader/reducer';
 import caseSelectReducer from '../../../app/reader/CaseSelect/CaseSelectReducer';
 import { annotationLayerReducer } from '../../../app/reader/AnnotationLayer/AnnotationLayerReducer';
-import documentsReducer from '../../../app/reader/DocumentList/DocumentsReducer';
+import documentsReducer from '../../../app/reader/Documents/DocumentsReducer';
 import PdfJsStub, { PAGE_WIDTH, PAGE_HEIGHT } from '../../helpers/PdfJsStub';
-import { onReceiveDocs } from '../../../app/reader/LoadingScreen/LoadingScreenActions';
+import { onReceiveDocs } from '../../../app/reader/Documents/DocumentsActions';
 import { onReceiveAnnotations } from '../../../app/reader/AnnotationLayer/AnnotationActions';
 
 import sinon from 'sinon';
